moving test suite to follow elizaos standards for publishing
