+++ packages/registry/index.json
@@ -13,7 +13,7 @@
             "version": "0.1.0",
             "runtimeVersion": "1.4.2",
             "platform": "node",
-            "publishedAt": "2025-08-13T20:19:58.093Z",
+            "publishedAt": "2025-08-13T23:13:01.557Z",
             "published": false
           }
         }
